@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections;
+using System.Collections.ObjectModel;
 using System.Data;
 using System.Data.Common;
+using System.Diagnostics;
 using System.IO;
 using System.Runtime.CompilerServices;
 using System.Threading;
@@ -11,7 +13,7 @@ namespace Dapper
 {
     internal sealed class DisposedReader : DbDataReader
     {
-        internal static readonly DisposedReader Instance = new DisposedReader();
+        internal static readonly DisposedReader Instance = new();
         private DisposedReader() { }
         public override int Depth => 0;
         public override int FieldCount => 0;
@@ -76,22 +78,11 @@ namespace Dapper
         public override object this[string name] => ThrowDisposed<object>();
     }
 
-    internal static class WrappedReader
+    internal sealed class DbWrappedReader : DbDataReader, IWrappedDataReader
     {
         // the purpose of wrapping here is to allow closing a reader to *also* close
         // the command, without having to explicitly hand the command back to the
-        // caller; what that actually looks like depends on what we get: if we are
-        // given a DbDataReader, we will surface a DbDataReader; if we are given
-        // a raw IDataReader, we will surface that; and if null: null
-        public static IDataReader Create(IDbCommand cmd, IDataReader reader)
-        {
-            if (cmd == null) return reader; // no need to wrap if no command
-
-            if (reader is DbDataReader dbr) return new DbWrappedReader(cmd, dbr);
-            if (reader != null) return new BasicWrappedReader(cmd, reader);
-            cmd.Dispose();
-            return null; // GIGO
-        }
+        // caller
         public static DbDataReader Create(IDbCommand cmd, DbDataReader reader)
         {
             if (cmd == null) return reader; // no need to wrap if no command
@@ -100,9 +91,7 @@ namespace Dapper
             cmd.Dispose();
             return null; // GIGO
         }
-    }
-    internal sealed class DbWrappedReader : DbDataReader, IWrappedDataReader
-    {
+
         private DbDataReader _reader;
         private IDbCommand _cmd;
 
@@ -210,97 +199,168 @@ namespace Dapper
         public override Task<bool> NextResultAsync(CancellationToken cancellationToken) => _reader.NextResultAsync(cancellationToken);
         public override Task<bool> ReadAsync(CancellationToken cancellationToken) => _reader.ReadAsync(cancellationToken);
         public override int VisibleFieldCount => _reader.VisibleFieldCount;
-        protected override DbDataReader GetDbDataReader(int ordinal) => (((IDataReader)_reader).GetData(ordinal) as DbDataReader) ?? throw new NotSupportedException();
+        protected override DbDataReader GetDbDataReader(int ordinal) => _reader.GetData(ordinal);
+
+#if NET5_0_OR_GREATER
+        public override Task CloseAsync() => _reader.CloseAsync();
+
+        public override ValueTask DisposeAsync() => _reader.DisposeAsync();
+
+        public override Task<ReadOnlyCollection<DbColumn>> GetColumnSchemaAsync(CancellationToken cancellationToken = default) => _reader.GetColumnSchemaAsync(cancellationToken);
+
+        public override Task<DataTable> GetSchemaTableAsync(CancellationToken cancellationToken = default) => base.GetSchemaTableAsync(cancellationToken);
+#endif
     }
 
-    internal class BasicWrappedReader : IWrappedDataReader
+    internal sealed class WrappedBasicReader : DbDataReader
     {
         private IDataReader _reader;
-        private IDbCommand _cmd;
 
-        IDataReader IWrappedDataReader.Reader => _reader;
-
-        IDbCommand IWrappedDataReader.Command => _cmd;
-
-        public BasicWrappedReader(IDbCommand cmd, IDataReader reader)
+        public WrappedBasicReader(IDataReader reader)
         {
-            _cmd = cmd;
-            _reader = reader;
+            Debug.Assert(reader is not DbDataReader); // or we wouldn't be here!
+            _reader = reader ?? throw new ArgumentNullException(nameof(reader));
         }
 
-        void IDataReader.Close() => _reader.Close();
+        public override bool HasRows => true; // have to assume that we do
+        public override void Close() => _reader.Close();
+        public override DataTable GetSchemaTable() => _reader.GetSchemaTable();
 
-        int IDataReader.Depth => _reader.Depth;
+#if PLAT_NO_REMOTING
+        [Obsolete("This Remoting API is not supported and throws PlatformNotSupportedException.", DiagnosticId = "SYSLIB0010", UrlFormat = "https://aka.ms/dotnet-warnings/{0}")]
+#endif
+        public override object InitializeLifetimeService() => throw new NotSupportedException();
 
-        DataTable IDataReader.GetSchemaTable() => _reader.GetSchemaTable();
+        public override int Depth => _reader.Depth;
 
-        bool IDataReader.IsClosed => _reader.IsClosed;
+        public override bool IsClosed => _reader.IsClosed;
 
-        bool IDataReader.NextResult() => _reader.NextResult();
+        public override bool NextResult() => _reader.NextResult();
 
-        bool IDataReader.Read() => _reader.Read();
+        public override bool Read() => _reader.Read();
 
-        int IDataReader.RecordsAffected => _reader.RecordsAffected;
+        public override int RecordsAffected => _reader.RecordsAffected;
 
-        void IDisposable.Dispose()
+        protected override void Dispose(bool disposing)
         {
-            _reader.Close();
-            _reader.Dispose();
-            _reader = DisposedReader.Instance;
-            _cmd?.Dispose();
-            _cmd = null;
+            if (disposing)
+            {
+                _reader.Close();
+                _reader.Dispose();
+                _reader = DisposedReader.Instance; // all future ops are no-ops
+            }
         }
 
-        int IDataRecord.FieldCount => _reader.FieldCount;
+        public override int FieldCount => _reader.FieldCount;
 
-        bool IDataRecord.GetBoolean(int i) => _reader.GetBoolean(i);
+        public override bool GetBoolean(int i) => _reader.GetBoolean(i);
 
-        byte IDataRecord.GetByte(int i) => _reader.GetByte(i);
+        public override byte GetByte(int i) => _reader.GetByte(i);
 
-        long IDataRecord.GetBytes(int i, long fieldOffset, byte[] buffer, int bufferoffset, int length) =>
+        public override long GetBytes(int i, long fieldOffset, byte[] buffer, int bufferoffset, int length) =>
             _reader.GetBytes(i, fieldOffset, buffer, bufferoffset, length);
 
-        char IDataRecord.GetChar(int i) => _reader.GetChar(i);
+        public override char GetChar(int i) => _reader.GetChar(i);
 
-        long IDataRecord.GetChars(int i, long fieldoffset, char[] buffer, int bufferoffset, int length) =>
+        public override long GetChars(int i, long fieldoffset, char[] buffer, int bufferoffset, int length) =>
             _reader.GetChars(i, fieldoffset, buffer, bufferoffset, length);
 
-        IDataReader IDataRecord.GetData(int i) => _reader.GetData(i);
+        public override string GetDataTypeName(int i) => _reader.GetDataTypeName(i);
 
-        string IDataRecord.GetDataTypeName(int i) => _reader.GetDataTypeName(i);
+        public override DateTime GetDateTime(int i) => _reader.GetDateTime(i);
 
-        DateTime IDataRecord.GetDateTime(int i) => _reader.GetDateTime(i);
+        public override decimal GetDecimal(int i) => _reader.GetDecimal(i);
 
-        decimal IDataRecord.GetDecimal(int i) => _reader.GetDecimal(i);
+        public override double GetDouble(int i) => _reader.GetDouble(i);
 
-        double IDataRecord.GetDouble(int i) => _reader.GetDouble(i);
+        public override Type GetFieldType(int i) => _reader.GetFieldType(i);
 
-        Type IDataRecord.GetFieldType(int i) => _reader.GetFieldType(i);
+        public override float GetFloat(int i) => _reader.GetFloat(i);
 
-        float IDataRecord.GetFloat(int i) => _reader.GetFloat(i);
+        public override Guid GetGuid(int i) => _reader.GetGuid(i);
 
-        Guid IDataRecord.GetGuid(int i) => _reader.GetGuid(i);
+        public override short GetInt16(int i) => _reader.GetInt16(i);
 
-        short IDataRecord.GetInt16(int i) => _reader.GetInt16(i);
+        public override int GetInt32(int i) => _reader.GetInt32(i);
 
-        int IDataRecord.GetInt32(int i) => _reader.GetInt32(i);
+        public override long GetInt64(int i) => _reader.GetInt64(i);
 
-        long IDataRecord.GetInt64(int i) => _reader.GetInt64(i);
+        public override string GetName(int i) => _reader.GetName(i);
 
-        string IDataRecord.GetName(int i) => _reader.GetName(i);
+        public override int GetOrdinal(string name) => _reader.GetOrdinal(name);
 
-        int IDataRecord.GetOrdinal(string name) => _reader.GetOrdinal(name);
+        public override string GetString(int i) => _reader.GetString(i);
 
-        string IDataRecord.GetString(int i) => _reader.GetString(i);
+        public override object GetValue(int i) => _reader.GetValue(i);
 
-        object IDataRecord.GetValue(int i) => _reader.GetValue(i);
+        public override int GetValues(object[] values) => _reader.GetValues(values);
 
-        int IDataRecord.GetValues(object[] values) => _reader.GetValues(values);
+        public override bool IsDBNull(int i) => _reader.IsDBNull(i);
 
-        bool IDataRecord.IsDBNull(int i) => _reader.IsDBNull(i);
+        public override object this[string name] => _reader[name];
 
-        object IDataRecord.this[string name] => _reader[name];
+        public override object this[int i] => _reader[i];
 
-        object IDataRecord.this[int i] => _reader[i];
+        public override T GetFieldValue<T>(int ordinal)
+        {
+            var value = _reader.GetValue(ordinal);
+            if (value is DBNull)
+            {
+                value = null;
+            }
+            return (T)value;
+        }
+        public override Task<T> GetFieldValueAsync<T>(int ordinal, CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            return Task.FromResult(GetFieldValue<T>(ordinal));
+        }
+        public override IEnumerator GetEnumerator() => _reader is IEnumerable e ? e.GetEnumerator()
+            : throw new NotImplementedException();
+        public override Type GetProviderSpecificFieldType(int ordinal) => _reader.GetFieldType(ordinal);
+        public override object GetProviderSpecificValue(int ordinal) => _reader.GetValue(ordinal);
+        public override int GetProviderSpecificValues(object[] values) => _reader.GetValues(values);
+        public override Stream GetStream(int ordinal) => throw new NotSupportedException();
+        public override TextReader GetTextReader(int ordinal) => throw new NotSupportedException();
+        public override Task<bool> IsDBNullAsync(int ordinal, CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            return Task.FromResult(_reader.IsDBNull(ordinal));
+        }
+        public override Task<bool> NextResultAsync(CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            return Task.FromResult(_reader.NextResult());
+        }
+        public override Task<bool> ReadAsync(CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            return Task.FromResult(_reader.Read());
+        }
+        public override int VisibleFieldCount => _reader.FieldCount;
+        protected override DbDataReader GetDbDataReader(int ordinal) => throw new NotSupportedException();
+
+#if NET5_0_OR_GREATER
+        public override Task CloseAsync()
+        {
+            _reader.Close();
+            return Task.CompletedTask;
+        }
+
+        public override ValueTask DisposeAsync()
+        {
+            _reader.Dispose();
+            return default;
+        }
+
+        public override Task<ReadOnlyCollection<DbColumn>> GetColumnSchemaAsync(CancellationToken cancellationToken = default)
+            => throw new NotSupportedException();
+
+        public override Task<DataTable> GetSchemaTableAsync(CancellationToken cancellationToken = default)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            return Task.FromResult(_reader.GetSchemaTable());
+        }
+#endif
     }
 }
